+++ src/main/java/tech/sirwellington/alchemy/http/AlchemyRequest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sir.wellington.alchemy.http;
+package tech.sirwellington.alchemy.http;
 
 import com.google.common.base.Joiner;
 import static com.google.common.collect.Lists.newArrayList;
@@ -27,7 +27,7 @@ import tech.sirwellington.alchemy.annotations.arguments.NonNull;
 import static sir.wellington.alchemy.arguments.Arguments.checkThat;
 import static sir.wellington.alchemy.arguments.Assertions.nonEmptyString;
 import static sir.wellington.alchemy.arguments.Assertions.notNull;
-import sir.wellington.alchemy.http.exceptions.AlchemyHttpException;
+import tech.sirwellington.alchemy.http.exceptions.AlchemyHttpException;
 
 /**
  *
