+++ src/main/java/tech/sirwellington/alchemy/http/HttpVerb.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sir.wellington.alchemy.http;
+package tech.sirwellington.alchemy.http;
 
 import static com.google.common.base.Charsets.UTF_8;
 import com.google.common.base.Strings;
@@ -45,10 +45,10 @@ import sir.wellington.alchemy.arguments.AlchemyAssertion;
 import static sir.wellington.alchemy.arguments.Assertions.nonEmptyString;
 import static sir.wellington.alchemy.arguments.Assertions.notNull;
 import sir.wellington.alchemy.arguments.FailedAssertionException;
-import sir.wellington.alchemy.http.HttpResponse;
-import sir.wellington.alchemy.http.exceptions.AlchemyHttpException;
-import sir.wellington.alchemy.http.exceptions.JsonException;
-import sir.wellington.alchemy.http.exceptions.OperationFailedException;
+import tech.sirwellington.alchemy.http.HttpResponse;
+import tech.sirwellington.alchemy.http.exceptions.AlchemyHttpException;
+import tech.sirwellington.alchemy.http.exceptions.JsonException;
+import tech.sirwellington.alchemy.http.exceptions.OperationFailedException;
 
 /**
  *
