+++ src/main/java/tech/sirwellington/alchemy/http/exceptions/OperationFailedException.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sir.wellington.alchemy.http.exceptions;
+package tech.sirwellington.alchemy.http.exceptions;
 
-import sir.wellington.alchemy.http.HttpResponse;
-import sir.wellington.alchemy.http.HttpRequest;
+import tech.sirwellington.alchemy.http.HttpResponse;
+import tech.sirwellington.alchemy.http.HttpRequest;
 
 /**
  *
